controllers and models files renamed topics.controllers.js and topics.models.js to be more descriptived
